+++ route/user_route/route.go
@@ -5,7 +5,7 @@ import (
 	"github.com/labstack/echo/v4/middleware"
 	"go-vote/handler/user_handler"
 	"go-vote/infra"
-	"go-vote/util/jwt"
+	middleware2 "go-vote/middleware"
 )
 
 const (
@@ -19,10 +19,10 @@ func Init(e *echo.Echo, inf *infra.Infra) {
 	// ~/user group
 	user := e.Group(groupUser)
 	user.POST("/register", user_handler.Register)
-	user.GET("/profile/:id", user_handler.GetProfile, middleware.JWTWithConfig(jwt.GetAccessConfig()))
+	user.GET("/profile/:id", user_handler.GetProfile, middleware.JWTWithConfig(middleware2.GetAccessConfig()))
 
 	// ~/user/auth group
 	auth := user.Group(groupAuth)
 	auth.POST("/login", user_handler.Login)
-	auth.POST("/refresh", user_handler.Refresh, middleware.JWTWithConfig(jwt.GetRefreshConfig()))
+	auth.POST("/refresh", user_handler.Refresh, middleware.JWTWithConfig(middleware2.GetRefreshConfig()))
 }
